refactor formPage and add Forms
